Fix templating bug that prevents the section list pages from being rendered
It turns out that it is incorrect to add a comment (or any other markup
I presume) outside the {{ define ... }} block. This fails silently, and
the generated template ends up being empty. This leads to empty files
for the section list pages, which is unpleasant.

By moving the comment to within the {{ define ... }} block this issue is
resolved.